app/requirements.txt
dj-database-url
django
git+https://github.com/matiasilva/django-ucamwebauth@master
google-auth
google-cloud-storage
gunicorn
psycopg2-binary
requests
simpleanalytics
whitenoise
django-s3-storage
django_ses
qrcode
Pillow